images: fix bug if iSCSI packages left blank
If iSCSI packages is left blank, the Dockerfile's verify packages step
may produce code that ends in a trailing `\` character, which will cause
the build to fail. Fix this by adding a final `&& echo ...` message.
`/bin/true` will work just as well, but the echo may be more useful for
debugging.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>