github: Added issue and pull request templates
This adds issue and pull request templates so the user will be asked for
general information, such as Ceph version, OS, kernel, to help make it
faster to debug issues and/or see if something is missing in a PR (e.g.
documentation updates).
The pull request template gives the user a checklist for certain points
that need to be kept in mind.

Signed-off-by: Alexander Trost <galexrt@googlemail.com>