contrib: wait for arm images to build the manifest
Before running the manifest command we must make sure that both jobs
(arm64 and amd64) are done. Currently it appears that the arm64 job take
more time to run than the amd64 one. And since the manifest command runs
on amd64 it fails. So now we wait for arm64 to be ready before running
the manifest command.

Signed-off-by: Sébastien Han <seb@redhat.com>