disk_list: backward compatibility with old ceph-disk
Older version of ceph-disk (when bluestore wasn't around) was creating
the lockbox partition for dmcrypt on partition number 3. With bluestore
this partition is now on position number 5. So we need to test if this
partition exists, if it does we use it.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1609007
Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit b7dd3bf1a66e6862aa81a992d9ff76c1bcb5cd3c)
Signed-off-by: Sébastien Han <seb@redhat.com>