rgw: read config startup from ceph.conf
It's easier to read rgw config options from the ceph.conf instead of
overriding it on the CLI.
This is a disruptive change for users that were relying on options like
RGW_CIVETWEB_IP and RGW_CIVETWEB_PORT. These options will have to
addressed in the ceph.conf so we don't break anything.

The options mentioned above are still active for the nano/demo scenario
which is really unique and not intended for production deployment.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1582411
Signed-off-by: Sébastien Han <seb@redhat.com>