contrib: build/push ceph-tagged images for Rook
The Rook project would like to have a standard way of stabilizing on a
specific Ceph package release. To accomplish this, add a script to
contrib that will build and push daemon-base images which are tagged
with the version of Ceph installed in them. Images are indended for Rook
and therefore CentOS-based.

The script enforces parity between Ceph versions available on the Ceph
package repository and the 'ceph' library on the Docker Hub. It looks
for all the Ceph package versions for a particular release or releases
(luminous and mimic to start), and it builds a Ceph image for each
package version which does not yet exist.

The first run of this script will likely be rather long, but subsequent
runs should build only 2 images at a time. The script is intended to be
able to be run as a regular cron job (daily? weekly?) that will pick up
the latest Ceph package version and publish it to the Ceph library on
Docker Hub without manual intervention.

The script also updates images if the centos base image is updated.
Builds when a centos base image is updated will again take a long time.

Images are versioned with v<cephversion>-<buildnumber> where cephversion
is the point release of Ceph (e.g. 12.2.2), and buildnumber starts at
zero and is incremented each time the image is rebuild due to base image
updates.

Images are tagged based on arch:
ceph/ceph-amd64:v<cephversion>-<buildnumber>
ceph/ceph-arm64:v<cephversion>-<buildnumber>

Manifest images are created that point to the x86_64 and aarch64 builds
if and only if both arches have existing images. Otherwise, no manifest
image is built. Manifest images are tagged without an arch, and the
build number for manifest images is the addition of the build numbers
for the x86_64 and aarch64 images:
ceph/ceph:v<cephversion>-<amd64-buildnumber+arm64-buildnumber>

Example:
Ceph version 12.2.2
The CentOS image for x86_64 has been updated twice since the first build
of the Ceph image. The ARM64 image has been updated three times since.
The resulting images will be:
  ceph/ceph-amd64:v12.2.2-2
  ceph/ceph-arm64:v12.2.2-3
  ceph/ceph:v12.2.2-5

The only manual intervention needed should be when a new Ceph version
is built and ready in ceph-container; at that point, the
`FLAVORS_TO_BUILD` variable in this script should be updated.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>