variable: do not test for fqdn if not necessary
If the shortname is returned by both commands (uname and the content of
/proc/sys/kernel/hostname) we don't need to display that we found an
FQDN config because it's not true. Enter, modify and display a message
only if they are different. It's fair to assume that an FQDN config is
present if both content are different.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1603202
Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit fa58105eb5d8765bf66ec1d2f52b2585cbcac7f4)
Signed-off-by: Sébastien Han <seb@redhat.com>