Add new 'ceph' base image
Instead of minimizing the 'daemon-base' image, establish a new 'ceph'
image which contains only Ceph binaries on top of the base os image. The
'daemon-base' image now builds iSCSI, NFS-Ganesha, and a few other
packages on top of the new 'ceph' image to build the same 'daemon-base'
image as before.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>