contrib: build/push ceph version tagged images
The Rook project would like to have a standard way of stabilizing on a
specific Ceph package release. To accomplish this, add a script to
contrib that will build and push daemon-base images which are tagged
with the version of Ceph installed in them. Images are indended for Rook
and therefore CentOS-based, but the script should support other distros
fairly easily.

The script enforces parity between Ceph versions available on the Ceph
package repository and the 'ceph' library on the Docker Hub. It looks
for all the Ceph package versions for a particular release or releases
(luminous and mimic to start), and it builds a Ceph image for each
package version which does not yet exist.

The first run of this script will likely be rather long, but subsequent
runs should build only 2 images at a time. The script is intended to be
able to be run as a regular cron job (daily? weekly?) that will pick up
the latest Ceph package version and publish it to the Ceph library on
Docker Hub without manual intervention.

The script also updates images if the base image is updated.
Builds when a base image is updated will again take a long time.

Images are versioned with v<cephversion>-<date> where cephversion
is the point release of Ceph (e.g. 12.2.2), and date is the current date
in YYYYMMDD format. A new image with a new date is created each time the
image is rebuilt due to base image updates.

Images are tagged based on arch:
ceph/ceph-amd64:v<cephversion>-<date>
ceph/ceph-arm64:v<cephversion>-<date>

Manifest images are created that point to the x86_64 and aarch64 builds
if and only if both arches have existing images. Otherwise, no manifest
image is built. Manifest images are tagged without an arch, and the
build number is also the date on which the manifest was created

The only manual intervention needed should be when a new Ceph version
is built and ready in ceph-container; at that point, the
`FLAVORS_TO_BUILD` variables in this script should be updated.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>