docker_exec: Reworking SIGTERM handler
The actual SIGTERM handler was racy as it didn't waited the targeted PID to actually die before running the post script.

Adding a wait() just after the kill was an initial though but failed as we are in a sig context and another wait is already pending in the exec().

This part of the program have to handle two cases :
- managing a fault of the exec'd program (return < 0 or SEGFAULT)
- managing the graceful shutdown triggered by the SIGTERM sent by docker

Both cases are considering watching the state of the exec'd program.
Having two wait() calls (one in _term(), one in exec()) isn't possible.

The actual implementation is considering the following :
- we just track the presence of the exec'd process by watching if the PID is running. The return code have no interest here.
- if we got a SIGTERM then we execute the pre & post scripts
- if we got any other signals like SIGCHLD, SIGINT, SIGBUS, ... then we make a best effort teardown as we cannot guarantee the behavior of the container

This patch make a single function (teardown()) to handle the process of shutting down the container.
SIGTERM (a docker stop) or a SIGCHLD (a process died) will lead to almost the same code path with a lot of tracing to get the reason why the container is dying.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1624341

Signed-off-by: Erwan Velu <erwan@redhat.com>