osd: Don't trim findmnt output
When sigterm_cleanup_post is called, it tried to umount OSD's filesystem.
To do so, it calls findmnt and try to umount it.

The issue in the actual code is that findmnt are :
- it trims unconditionnaly two lines
- it doesn't consider the virtual system case

Removing a fixed set of lines could be wrong as some configurations could report a single output.
This over-filtering will in that case report nothing & no umount.

This patch offer to switch to an explicit filtering for searching block devices that have a full path as findmnt reports some virtual filesystems like sysfs or tmpfs.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1624341

Signed-off-by: Erwan Velu <erwan@redhat.com>