contrib: fix stable release build
This fixes the situation where we have a tag v3.0.1 but v3.0.1rc1 exists
too. The old version of the grep will match both to true so it won't
find any new addition. The regex wasn't precise enough. Now we either
check the exact match is the last or in-between the list.

Signed-off-by: Sébastien Han <seb@redhat.com>