centos: Adding CEPH_VERSION=master support
Testing the master release of ceph is useful for many.
This patch is about adding a new flavor to support it.

To achieve that the following changes are made in this commit :
- download.ceph.com remains the repo for releases
- shaman.ceph.com   becomes the repo for master
REPO_URL variable is feed with the proper repo regarding the CEPH_VERSION
Finding the right repo on shaman requires 'jq' which is provided by epel.
So epel is now installed _before_ we try to install the ceph/ganesha rpms.

Note both ganesha & ceph are taken from shaman when CEPH_VERSION is set to master

On shaman, the main package is name ceph-release-1-0 while its ceph-release-1-1 on download.ceph.com
RELEASE_VER variable is feed with the appropriate value regarding the CEPH_VERSION

As shaman doesn't provides any arm64 builds, the master build is not
included in AARCH64_FLAVORS_TO_BUILD

To use that feature:

    make FLAVORS=master,centos,7 build

Fixes: https://github.com/ceph/ceph-container/issues/1200

Signed-off-by: Erwan Velu <erwan@redhat.com>