mon: create bootstrap keys in //
This will speed up the bootstrap of the initial monitor since we are now
using a loop and pile up all the creations and then wait for them to be
complete.
Also, there is no need to check for the existence of the keys since this
code is only triggered when a new monitor is detected.

Signed-off-by: Sébastien Han <seb@redhat.com>