contrib: fix error handling in ceph/ceph scripts
Add `export SHELLOPTS` to export the parent's shell options to
subshells so they definitely run with `-e` option.

Add `-E` (errtrace) option to scripts since this is needed for `-e` to
properly fail from subshells.

Move ERR trap to build config so that the manifest creation script also
gets this.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>