nano: add static sree
We don't have a package yet, the code hasn't changed for months, the
tarball is under 500KB so let's add it to the daemon image.
This avoids downloading the tar each time we boot a new cn container and
also brings back the ability to work offline or behind a proxy.

Closes: https://github.com/ceph/cn/issues/72
Signed-off-by: Sébastien Han <seb@redhat.com>