docker_exec: Enabling staylive from teardown()
When a SIGCHLD is triggers, the stayalive feature wasn't called.

This patch is about making the stayalive being called directly from the
teardown() code.

The rework done by 78d9106a1144ef902a1c0cb81a0d348775b25023 gives an
easy way to enable the stayalive as the nearly end of the teardown code.

This patch add ERR as part of the trapped signals.

Note this patch is also reworking the way the container ID is caught by
extracting it from the cgroup info instead of the hostname.

Signed-off-by: Erwan Velu <erwan@redhat.com>