src/daemon/demo.sh: Dynamic sleep instead of harcoded value
The current code was waiting for 5 secs before trying to create the bucket.

The whole container usually starts in 25 secs meaning that a 1/5th is spent here.

This patch is poking the s3 service to see if it's available.
The loop is waiting up to 5 seconds.

Once the radosgw is ready, the bucket is created.

Tests on a private server showed that a single loop is enough, saving 5 seconds on the start process.

Signed-off-by: Erwan Velu <erwan@redhat.com>