centos: Enabling CR repository on Centos
Testing on master could trigger some racing between the specfile descriptions and the state of the Centos repository.

The Travis CI ran and reported an issue like :

    Error: Package: 2:ceph-selinux-14.0.1-1009.g1756025.el7.x86_64 (Ceph)
           Requires: selinux-policy-base >= 3.13.1-229.el7_6.5
           Available: selinux-policy-minimum-3.13.1-192.el7.noarch (base)

The -229.el7_6.5 version of the package is required by ceph-selinux but is not yet available on the update repository.

The CR repository (https://wiki.centos.org/AdditionalResources/Repositories/CR) can be used to reach the packages before they get into the update repo.
    "[CR] makes generally available packages that will appear in the next point release of CentOS"

This commit is about enabling CR so yum can reach the latest version of the packages.

Signed-off-by: Erwan Velu <erwan@redhat.com>