contrib/build-push-ceph-container-imgs.sh: Disabling ARM builds
ARM builds are pretty much slower & weaker than x86 ones.
There is situations where we need to disable the ARM builds not to block the x86 ones.
When this situation occurs, the BUILD_ARM variable can be emptied.

Once the ARM build are stable again, this variable can be set to anything so it is reactivated.

Note this patch is disabling the ARM builds are we are currently facing an issue with them.

Signed-off-by: Erwan Velu <evelu@redhat.com>