daemon: Tuning ceph's memory when container starts
This commit is about introducing a dynamic tuning of ceph's.
The concept is to collect first the amount of available memory reported by the system.

If the cgroup report a reasonable value, this value is used, unless (if cgroups reports 8Exabytes), the value is taken from the system.

This value is then reduced from a couple of MB to let some room for the other process to run.
If this value exceed 4GB, the amount of memory is caped to avoid giving all possible memory to ceph.

Then, the 3 ceph metrics are computed :

    +----------------------+ -> max system memory
    |                      |
    |                      |
    |----------------------| -> available memory
    |                      |
    |----------------------| -> available memory - 50MB
    | osd_memory_target    | -> memory used by ceph at most
    |--------------------- |
    |                      |
    |                      |
    |                      |
    |                      |
    |                      |
    |----------------------|
    | osd_memory_cache_min | -> minimum size of memory cache
    |----------------------|
    |                      |
    |                      |
    |                      |
    |                      |
    |                      |
    |                      |
    |----------------------|
    | osd_memory_base      | -> "base" memory that an OSD will need once it's reached steady state"
    |----------------------|
    |                      |
    |                      |
    |                      |
    |                      |
    |                      |
    |                      |
    +----------------------+ -> Bottom of address space

Once computed, and only if the metrics have decent values (too small value would be rejected), the metrics are injected inside the ceph.conf prior the daemon starting.

A typical output for a 512MB cluster :
2018-12-19 14:32:40  /opt/ceph-container/bin/entrypoint.sh: Found 500MB of available memory (525000704 bytes)
2018-12-19 14:32:40  /opt/ceph-container/bin/entrypoint.sh: Tuning memory : osd_memory_base=250MB, osd_memory_cache_min=350MB, osd_memory_target=450MB

[...]
The /etc/ceph/ceph.conf looks like below where osd_memory_target, osd_memory_base, osd_memory_cache_min are set accordingly.

    [root@ceph-nano-memory-faa32aebf00b /]# cat /etc/ceph/ceph.conf
        [global]
        fsid = 0413b45e-4fea-4672-af1b-879eb5979a23
        mon initial members = ceph-nano-memory-faa32aebf00b
        mon host = 127.0.0.1
        osd crush chooseleaf type = 0
        osd journal size = 100
        public network = 0.0.0.0/0
        cluster network = 0.0.0.0/0
        log file = /dev/null
        osd pool default size = 1
        osd_memory_target = 472571904
        osd_memory_base = 262500352
        osd_memory_cache_min = 367536128
        [...]

Signed-off-by: Erwan Velu <erwan@redhat.com>