contrib: Add alpha/beta tag filtering
We currently only filter image tags that are like:
  - vX.X.X
  - vX.X.XrcX

Unfortunately we have a v4.0.0beta tag present on docker hub. Due
to the regex this tag is considered as v4.0.0 so we can build this
tag.
This commit update the tags regex and add alpha and beta suffix like
for rc.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>