rhel8: rm ubi.repo
The RHEL 8's UBI (universal base image) has a ubi.repo file that points
at repositories that are only available with subscription-manager.

If these repositories are not available, dnf will exit with an error
(rhbz#1701290). We currently build the Ceph RHEL images in an
environment that does not use subscription-manager, so dnf cannot access
the UBI content, and it bails out at this point.

We already get RHEL 8 build content dynamically from OSBS's Yum repository
injection feature, so we do not need anything in this ubi.repo file.

Delete the ubi.repo file entirely to avoid the errors associated with
the repositories here.

UBI is a new initiative, so maybe we will find a better workaround in
the future.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>