osd_disk_activate: allow to disable ceph-osd start
In some case, we don't want to activate and start the OSD process in
the same entrypoint.
In order to allocate an OSD id on prepared OSD, we need to activate it
without start it.
That way, we're able to get the OSD id for systemd unit script.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1670734

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>
(cherry picked from commit 3bef95ddf10aacad70cc0298a37073084d96191f)