daemon/start_mon.sh: add mon without port value
When bootstrapping an initial cluster, we don't need to add the mon
to the monmap via 'ceph mon add' command.
But when we want to scale up the monitors then we need it.
Currently the new mon added will be out of quorum because the addr
in the ceph config and in the monmap won't match.

start_mon(): ceph --cluster ceph mon add xxx 192.168.100.5:3300
mon.xxx at v2:192.168.100.5:3300/0 already exists

Even if the ceph conf uses v2+v1 then the new mon will have only v2
address configured in the monmap creating a conflict.

0 mon.xxx@0(probing) e3  monmap addrs for rank 4 changed, i am
  [v2:192.168.100.5:3300/0,v1:192.168.100.5:6789/0], monmap is
  v2:192.168.100.5:3300/0, respawning
0 mon.xxx@0(probing) e3 respawn

Removing the port value on the 'ceph mon add' command allows to add
a monitor with the msgr configuration defined in the ceph config.

This change is nautilus+ only.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1719013
Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1722066

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>