contrib: override releases list on stable branch
There's multiple issues with the build-push-ceph-container-imgs.sh
script:
  - we execute the script from master but checkout the code from stable
branch when building a new tag.
  - the ceph_releases list is static and should not be the same if we're
building on tag (TAGGED_HEAD=true) or not.

* stable-3.2: luminous and mimic
* stable-4.0: nautilus

Currently on stable-3.2 and stable-4.0 branches we always build
luminous, mimic and nautilus for arm64 architecture (because hardcoded
in the build-push-ceph-container-imgs-arm64.sh script) while we build
releases based on the FLAVORS value from the Makefile for x86_64.

This commit allows to override the ceph releases list when we're
building a new tag.
Because the new list won't have master then we need to change the
condition in the create_registry_manifest function (master was supposed
to be the first value).

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>