Introduce CI_CONTAINER to do mostly what DEVEL does
Slight differences in tagging, and build only daemon-base for
centos 7.  Don't rebuild for changes to the ceph-container source.
The intent is to use CI_CONTAINER builds for ceph-ci.git builds,
so that a container is present.  This script will be called from the
main ceph-dev-new build job with CI_CONTAINER set when appropriate.

Signed-off-by: Dan Mick <dan.mick@redhat.com>