cleanup: disable repos pointing to chacra
Repos on chacra are quite ephemeral, they last for 2 weeks only. If
someone is using an image that is 3 weeks old, the repos will be gone
and they won't be able to install any packages.
So disabling these repos is easy and less confusing for users.

They can always be re-enabled with:

yum-config-manager --enable ceph-iscsi ceph-iscsi-source ceph-iscsi-noarch tcmu-runner tcmu-runner-noarch tcmu-runner-source python-rtslib-source python-rtslib-noarch python-rtslib

Signed-off-by: Sébastien Han <seb@redhat.com>