exec: catch sigterm when bash does not notice it
Currently when using 'docker stop' to stop a container, docker send
SIGTERM to the PID 1 within the container.
Currently PID 1 is the entrypoint.sh as seen here:

```
[root@mon0 bin]# ps fauxwww
USER       PID %CPU %MEM    VSZ   RSS TTY      STAT START   TIME COMMAND
root      1900  0.0  0.0  11820  1976 ?        Ss   Aug21   0:00 bash
root     12676  0.0  0.0  51736  1728 ?        R+   10:38   0:00  \_ ps fauxwww
root         1  0.0  0.0  12080  1984 ?        Ss   Aug21   0:00 /bin/bash /opt/ceph-container/bin/entrypoint.sh
ceph       148  1.0 18.1 1149368 703164 ?      Sl   Aug21  32:39 /usr/bin/ceph-mon --cluster ceph --default-log-to-file=false --default-mon-cluster-log-to-file=false --setuser ceph --setgroup ceph -d --mon-cluster-log-to-stderr --log-stderr-prefix=debug  -i mon0 --mon-data /var/lib/ceph/mon/ceph-mon0 --public-addr 192.168.39.10
```

The problem is that docker_exec.sh which wraps bash's exec sub-command
positions a pid wait on the last process that was executed by the
entrypoint, in our case a ceph daemon process. So docker_exec runs the
ceph process in background and waits for it die.

The problem is that we trap on various calls and the only one that works
is ERR, ERR is an internal signal only known by bash. The reason why we
are getting ERR and not the SIGTERM sent by docker is because we trap on
the wrong PID.

So basically PID 1 receives SIGTERM, terminates and all the environment
goes away, since our wrapped exec function traps for ERR and we use set
-e, anything failing will fall on the ERR trap. That why we don't get
the desired behavior and nothing is cleaned up, functions like
sigterm_cleanup_pre are never calles.

The current attempt to fix the current behavior is to read the return
code of the child process that got killed, in our case it's 143 so this
corresponds to SIGTERM. If we detect that return code, we call our
cleanup functions.

Signed-off-by: Sébastien Han <seb@redhat.com>