exec: Remove ERR trap so other traps function
As far as I can tell, the ERR trap is overriding all other traps defined
here, so they are not being called.  Meanwhile the ERR trap seems to
only cover the few lines that follow it until wait happens, for as soon
as the trap is entered the ERR trap is unset.

As the ERR trap is only covering a few lines that should be completely
reliable, and isn't handling cleanup for any errors that occur before
`exec` is called, this should be safe to remove.

This allows further development of the traps at hand here, such as
passing SIGHUP through.

Signed-off-by: Sachi King <nakato@nakato.io>