osd_volume_activate: fix osd_volume_lvm function
90511a4e0231d42d98171ac456e714406605de15 introduced a bug in
osd_volume_lvm function.
OSDs couldn't start because `$CEPH_VOLUME_LIST_JSON` is seen as an empty
variable when consumed in this function.
Since there's no need to have a separate function for detecting the
volume type, let's refact this part of the code so we can fix this
issue.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>