src/daemon: tune OSD memory on demo
There's no reason to do the memory tunning on the MON nodes. This should
be done on the OSD nodes instead.
Also this process is done when bluestore is used.
Finally this process is only for demo.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>
(cherry picked from commit 9fe8369ef807b643499fa8b53b06ccfd5fc6f465)