ceph-build-config: fix infinite loop on registry
Now the curl command against the docker hub registry to list the
tags per page doesn't fail on non existing page.
Because there's only a condition on the curl command then there's an
infinite loop with the page number incrementing.
We can break the loop when the next attribute is null.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>