build: only build stable releases
Currently the regex builds every package, which leads to issue when not
building the stable minor '2' because some packages do not exist.
For example, we hit:

No package ceph-mgr-k8sevents-14.1.0 available.

It doesn't make sense to build non-stable packages because we have dev
builds.
So this will avoid packages issues as well as speeding up the build
time.

Signed-off-by: Sébastien Han <seb@redhat.com>