Fix activating legacy osd-disk partitioned nvme device
When the daemon tries to mount an nvme device which was prepared by ceph-disk, the partition path should be /dev/nvme0n1p1, the current osd_disk_activate.sh will use the hdd rule and try to find the partition as /dev/nvme0n11, which missed a 'p'.

As "dev_part" already been widly used, we can use it here to avoid un-reliable text concatenation.

Signed-off-by: Tang Jianyu <jianyu@jianyu.net>