tests: update docker insecure registry
The docker insecure registry setup should be done in the
/etc/docker/daemon.json file rather than /etc/sysconfig/docker.
We should also not block the docker.io registry because we're
using some container images from that registry for dashboard
intergration into ceph-ansible.
Finally we only need to set the ceph_docker_registry value into the
insecure registry value instead of using the public_network variable.
This was already done for podman.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>