ubi8: ignore absent pulp repositories
OSBS introduced a new "ignore_absent_pulp_repos" setting. This setting
tells ODCS to proceed when it cannot find some of the expected
repositories in pulp.

The reason we must enable this setting is that Red Hat release
engineering has not created all the pulp repositories in
content_sets.yml yet. We can revert this commit when all of those pulp
repositories are ready.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>