ceph-releases/octopus: only package ceph-mgr for centos/7
octopus is already a "python3 only" release, but centos7 is missing
quite a lot of python3 depdendencies for almost all of ceph-mgr-*
packages, like: ceph-mgr-rook and ceph-mgr-dashboard, the missing
dependencies are:

- python3-jsonpatch
- python3-routes
- python3-remoto
- python3-jinja2
- python3-jwt
- python3-kubernetes
- python3-scipy
- python3-cherrypy

since octopus built for RHEL/CentOS 7 is part of upgrade path, a fully
functional dashboard is not our priority, but we do need ceph-mgr is
critical. so let's just package ceph-mgr for octopus on el7.

Signed-off-by: Kefu Chai <kchai@redhat.com>