daemon: set root owner/group on etcdctl binary
When installing etcdctl binary via the upstream tarball the owner/group
is set to 478493:89939 which could lead to failure when using podman and
rootless containers.
There's no need to have such uid/gid on the etcdctl binary. When the
installation is done by package then the owner/group is root.

Closes: #1725

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>