contrib: do not overwrite images created for el8
before this change, we build octopus container images for both el7 and
el8, and tag them using the same set of tag names. so the last build
wins in the racing. the consumers of these images still expect the el8
builds by asking the images by names encoded using sha1.

after this change, the el7 container images are not tagged with names
without distro_release encoded, so the el7 images does not overwrite
the el8 ones even if they are built and pushed after el8 ones with
the same sha1 and branch names.

Fixes: https://tracker.ceph.com/issues/48162
Signed-off-by: Kefu Chai <tchaikov@gmail.com>