contrib/build-push-ceph-container-imgs.sh: undo kubic repo hack
There's a reasonable chance that podman 2 fixes the issue in
https://github.com/containers/libpod/issues/5306 (the bug
seems to imply that 1.9.1 has the fix), and the kubic
repo causes problems because of podman-catatonit, sometimes
called just "catatonit", and its dependencies on and of podman-docker.

Uninstall any remnants of the kubic repo, and switch back to
the container-tools:rhel8 module for podman etc.

Signed-off-by: Dan Mick <dmick@redhat.com>