contrib/build-push-ceph-container-images.sh: don't push short tags for arm
Currently, using normal manifests/tags as we are, the 'branch' and
'sha1' tags collide for x86_64 and arm64.  While the solution for this
is probably to create 'manifest lists' or 'fat manifests' for those
tags, and point to both archs, the way release containers do,
until we can get there (and modify prune-quay.py to handle the new
form of tags as well), this will allow us to build containers
with the full tag at least, and not tie the build speed of both
arch builders together.

Signed-off-by: Dan Mick <dmick@redhat.com>