centos-arm64/8: enable ceph-iscsi packages
Since we have tcmu-runner package available on shaman for arm64 which was
the last piece missing to be able to install ceph-iscsi, then we can
enable the ceph-iscsi installation on the arm64 architecture as well.
This is done only on CentOS 8 distribution.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>
(cherry picked from commit 4950db359b415dc01834482e458762f9aa1e82f9)