daemon-base/centos: simplify tcmu shaman repo code
Instead of doing two curl calls on shaman for retrieving the tcmu-runner
repository information then we can do it in a single one via the /latest
endpoint and the arch parameter (instead of the search API).
This also makes sure we're using the -o curl option to store the repository
information on disk instead of using output redirection.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>