centos: add "vault" binary to the image
Rook needs the vault binary to interact with a vault server. We cannot
add this to the Rook image since this is executed on Deployments
ressources which run on the host where the OSDs are running. This is not
a command we can run from the Operator itself so it should be in the
Ceph image.

It looks like the repo does not have any arm64 package so building for
x86 only.

Signed-off-by: Sébastien Han <seb@redhat.com>