contrib: build/push wait for arm build for stable
For tagged head release (like v6.0.4, v5.0.11, etc..) then we can wait
for the arm build to be finished before starting the manifest creation.
Because we only have one arm image to build, we won't have to wait so
much before retrieving that image from the amd64 job.
Before this change, the manifest was only created with the amd64 image
and then updated manually outside of the CI.
Let's avoid that.

Signed-off-by: Dimitri Savineau <dsavinea@redhat.com>