don't strip and gzexe binaries in CLEAN stage
This commit removes some attempts to make the final image smaller,
because they introduce other problems and have very little impact on the
overall image-size and just adds complexity.

Regarding `gzexe`:
gzexe is a tool that compresses binaries and wrap the
compressed data inside a shellscript. This will then automatically
decompress into a temporary file and execute that data whenever the
wrapper-shellscript is called.
ceph-container has been using this to compress `confd`, `kubectl`, and
`etcdctl` (and previously also `forego`). A critical problem here is that the
wrapper script cleans after itself with a subshell like this:

`(sleep 5; rm -fr "$gztmpdir") 2>/dev/null &`

The entrypoints/scenario-scrips sets traps on a list of signals, among
them SIGCHLD. So whenever the entrypoint has the desired scenario, if
the timing is right, the subshell of gzexe will exit, and thus trigger a
SIGCHLD and cause the entire docker image to shut down. There are a few
issues filed against this problem, when using etcd or confd such as:
https://github.com/ceph/ceph-container/issues/1443

Regarding `strip` of binaries:
There is a fatal bug in the bash-function `ifstrip` that's present in
the `__DOCKERFILE_CLEAN_*__` scripts, and has been there since 2018
atleast. This bug also prohibits any binaries to be stripped, becuase
the function `compgen -g "$1"` is invalid for all arguments we feed it,
and the function will never reach the `strip -s` stage.
The manual for `compgen` states that `-g` options is used to lookup
groups (from /etc/group)! And giving filenames to it, will result in an
error. For example:

```
[root@local /]# grep "^a" /etc/group
adm:x:4:
audio:x:63:
[root@local /]# compgen -g a
adm
audio
[root@local /]# compgen -g /usr/ls
[root@local /]# echo $?
1
```

The suspicion is that the original author intended to actually use
`compgen -G` (capital G). Which has the desired(?) usage. That option
will actually check if a globpattern expands to possible completions, so
that the bash-function ifstrip wouldn't call `strip -s` on nonexisting
files/directories. `compgen -G` seems more or less the same as using
`[[-f <filename>]]`, but works on globs without invoking a lot of
recursion/looping, so it's quite a clever trick. However, since this has
never worked due to the bug and no binaries has ever been stripped since
2018, lets remove it to simplify the cleanup-stage a bit.

During trials we found that the binaries we try to strip in he files
don't reduce much anyhow, only a few kilobytes (except for kubectl which
actually strips rather well). And the `confd` binary is (as of 0.16)
compressed with UPX already, and running `strip` on that will fail.

Signed-off-by: Björn Zettergren <bjorn.zettergren@deltaprojects.com>