base: add packages
Since we moved to stream 8, `procps-ng` and `hostname` packages
aren't present anymore. This adds them back in the image.
Having `ps` binary doesn't consume a lot of space and is
useful for devel/debug purposes.
(Although `hostname` is added back for ci/testing only, it doesn't hurt
to add it back too)

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>