use official centos repos for nfs-ganesha packages
Do not use buildlogs.centos.org as a repo. Those locations are not
guaranteed to be permanent storehouses for RPM builds, and they are
generally assumed to also have non-release versions as well.

Instead, use official Centos mirrors for nfs-ganesha packages starting
with ganesha v4. This commit also codifies the NFS repo for stream9,
which changes its URL structure significantly from stream8. Though
stream9 is not used yet, it will be coming soon, and it is important to
record the change in NFS repo that happens in stream9 alongside these
changes to make the upcoming transition easier.

To handle the additional complexity that now exists with ganesha repos
differing between stream8 and stream9, and also differing between Ceph
versions, this commit changes the old if,elsif,elsif,else approach to
setting the ganesha repo in the Dockerfile to one where __GANESHA_REPO__
is defined and can be overridden in Ceph version directories.

All builds using stream8 or stream9 begin with a base __GANESHA_REPO__
that allows the ceph version directory to specify __GANESHA_VER__. This
base __GANESHA_REPO__ should be suitable for main builds for the
forseeable future. It is also suitable for Ceph reef (which uses
__GANESHA_VER__ = 5) and Ceph quincy (which uses __GANESHA_VER__ = 4).
Ceph main builds obviously use the latest __GANESHA_VER__ = 5 value.

The above applies to the Ceph main, Reef, and Quincy builds which were
previously using buildlogs repos.

Ceph releases for Pacific and earlier include NFS-Ganesha packages that
are built alongside the Ceph RPMs which have specific API compatibility
options set. Therefore, __GANESHA_REPO__ is overridden for pacific and
lower builds and is set to the same content that was previously selected
via the if,elsif,elsif,else behavior.

Signed-off-by: Blaine Gardner <blaine.gardner@ibm.com>