Remove custom c9 __DOCKERFILE_INSTALL__; use generic version
The c9 version didn't get the prerelease changes for ceph-release,
so something was necessary to add them; besides that, the only
differences were the old method of selecting ganesha for later
releases and the enablement of crb instead of powertools in the
last line. Since we already had code to test centos8 in
daemon-base/__DOCKERFILE_INSTALL__, one more small conditional
won't hurt.

Also add the projectname (ceph) to the prerelease path (since
ceph-build's sync-push tool now adds it).  This affects
the __DOCKERFILE_INSTALL_ as well as contrib/ceph-build-config.sh

Signed-off-by: Dan Mick <dmick@redhat.com>